Move server.js and config.js to root
