+++ lib/controllers/applications.js
@@ -1,5 +1,5 @@
 const cdnServices = require('../services/cdn');
-const config = require('../config');
+const config = require('../../config');
 const Boom = require('@hapi/boom');
 
 module.exports = {
