+++ lib/controllers/deploy-sites.js
@@ -1,6 +1,6 @@
 const githubServices = require('../services/github');
 const releasesService = require('../services/releases');
-const config = require('../config');
+const config = require('../../config');
 const Boom = require('@hapi/boom');
 
 const PIX_SITE_REPO_NAME = 'pix-site';
