+++ lib/services/cdn.js
@@ -1,5 +1,5 @@
 const axios = require('axios');
-const config = require('../config');
+const config = require('../../config');
 const _ = require('lodash');
 
 const CDN_URL = 'https://console.baleen.cloud/api';
