+++ lib/services/github.js
@@ -1,6 +1,6 @@
 const { Octokit } = require('@octokit/rest');
 const axios = require('axios');
-const settings = require('../config');
+const settings = require('../../config');
 
 const color = {
   'team-evaluation': '#FDEEC1',
