+++ lib/services/google-sheet.js
@@ -1,6 +1,6 @@
 const axios = require('axios');
 const _ = require('lodash');
-const settings = require('../config');
+const settings = require('../../config');
 const URL_SPREADSHEET_API = 'https://sheets.googleapis.com/v4/spreadsheets/';
 const SPREADSHEET_VALUES = 'tips!A1:E500';
