+++ lib/services/releases.js
@@ -1,6 +1,6 @@
 const process = require('process');
 const util = require('util');
-const config = require('../config');
+const config = require('../../config');
 const ScalingoClient = require('./scalingo-client');
 const exec = util.promisify(require('child_process').exec);
