+++ lib/services/scalingo-client.js
@@ -1,5 +1,5 @@
 const scalingo = require('scalingo');
-const config = require('../config');
+const config = require('../../config');
 
 class ScalingoClient {
   constructor(client, environment) {
