+++ lib/services/sendinblue-report.js
@@ -1,7 +1,7 @@
 const axios = require('axios');
 const SibApiV3Sdk = require('sib-api-v3-sdk');
 const dayjs = require('dayjs');
-const config = require('../config');
+const config = require('../../config');
 
 function _instanciateSendInBlueTransactionalEmailsAPi() {
   const defaultClient = SibApiV3Sdk.ApiClient.instance;
