+++ lib/services/slack/security.js
@@ -2,7 +2,7 @@ const crypto = require('crypto');
 const querystring = require('querystring');
 const tsscmp = require('tsscmp');
 const Boom = require('@hapi/boom');
-const config = require('../../config');
+const config = require('../../../config');
 
 function verifyRequestSignature(signingSecret, body, signature, requestTimestamp) {
   if (signature === undefined || requestTimestamp === undefined) {
