+++ lib/services/slack/shortcuts.js
@@ -1,7 +1,7 @@
 const axios = require('axios');
 const publishReleaseTypeSelectionModal = require('./surfaces/modals/publish-release/release-type-selection');
 const deployReleaseTagSelectionModal = require('./surfaces/modals/deploy-release/release-tag-selection');
-const config = require('../../config');
+const config = require('../../../config');
 
 module.exports = {
