+++ lib/services/slack/surfaces/messages/post-message.js
@@ -1,5 +1,5 @@
 const axios = require('axios');
-const config = require('../../../../config');
+const config = require('../../../../../config');
 
 module.exports = (message) => {
   const options = {
