+++ test/unit/lib/services/sendinblue-report_test.js
@@ -1,7 +1,7 @@
 const { describe, it } = require('mocha');
 const { sinon } = require('../../../test-helper');
 
-const config = require('../../../../lib/config');
+const config = require('../../../../config');
 const axios = require('axios');
 const SibApiV3Sdk = require('sib-api-v3-sdk');
 const sendinblueReport = require('../../../../lib/services/sendinblue-report');
